course04-AOP/src/main/java/soundsystem/TrackCounter.java
package soundsystem;

import org.aspectj.lang.annotation.Aspect;
import org.aspectj.lang.annotation.Before;

import java.util.HashMap;
import java.util.Map;

/**
 * 记录每个磁道的播放数
 *
 * @author justZero
 * @since 2018/12/31
 */
@Aspect
public class TrackCounter {

    private Map<Integer, Integer> trackCounts = new HashMap<>();

    @Before("execution(* soundsystem.album.CompactDisc.playTrack(int))" +
            "&& args(trackNumber)")
    public void countTrack(int trackNumber) {
        int currentCount = getPlayCount(trackNumber);
        trackCounts.put(trackNumber, currentCount+1);
    }

    public int getPlayCount(int trackNumber) {
        return trackCounts.getOrDefault(trackNumber, 0);
    }

}
